extractors: ida: move os extraction to global module
